feat: r22 管理端待审批页签 + 声明占用 CSV 导出 + X-API-Token 只读访问（1.11.0）
新增能力
- 管理端「待审批」专页签：GET /api/admin/records?status=PENDING 仅返回待审批预约（不传 date 即不按日期过滤，覆盖跨天场次），前端可就地批准/拒绝；概览新增待审批计数卡片与页签徽标（T52）
- 声明占用 CSV 导出：GET /api/admin/asset-claims/export 输出带 BOM 的 CSV（资源编号/资源名称/声明次数/最近占用日期 + 合计行），只读（T53）
- X-API-Token 只读访问：GET 请求可携带 X-API-Token 头免 Cookie 认证，命中刷新 last_used_at；POST、令牌吊销与越权端点一律拒绝（T54）

修复（均为既有缺陷，本轮发现）
- records() 的 status 过滤写成字面量 '?'，条件恒不成立，待审批过滤实际不生效
- 令牌吊销 POST /api/me/tokens/{id}/revoke 误置于 GET 分支：GET 请求体不被解析、request_id 恒缺失，该端点恒返回 400、功能不可用（契约本已声明为 POST）
- admin.js 的声明占用表调用未定义的 stamp()，致该表始终加载失败
- utilization_export 表头 "利用率%" 的 % 未转义（-Wformat 告警且表头错乱）
- 编译告警清零：main.c 的 Config 初始化多余元素、metrics.c 两处 sign-compare、unit.c 的 Config 初始化缺字段——这些告警会让 build.ps1 在 $ErrorActionPreference='Stop' 下提前中断
- 清除 db.c/service.c 中 41 处 [mig]/[check]/[tc] 调试 fprintf(stderr,...)，它们污染 stderr 并同样会中断 build.ps1

文档与契约
- CONTRACT.md 补齐既有漂移字段（records 的 username/lab_id/note、events.id、utilization 实机时三列）并补入 r21 遗漏的 asset-claims；契约端点 36 → 39 且实现零漂移
- CHANGELOG/PROGRESS/README 同步；LAB_VERSION 1.10.0 → 1.11.0

验收
- build.ps1 全绿（三目标零告警 + 24 单元测试）
- 集成 55 项全绿（52 项既有断言未改动，仅新增 T52-T54）
- 契约 39 端点 + 14 错误场景全绿且零漂移

+++ src/metrics.c
@@ -45,10 +45,10 @@ char *metrics_prometheus(void){
  for(int i=0;i<METRIC_BUCKETS;i++){
   cum+=g_buckets[i];
   char le[24];if(i<METRIC_BUCKETS-1)snprintf(le,sizeof le,"%g",bucket_max[i]);else snprintf(le,sizeof le,"+Inf");
-  if((size_t)n+160>=(int)cap){cap*=2;t=realloc(t,(size_t)cap);if(!t)return NULL;}
+  if((size_t)n+160>=(size_t)cap){cap*=2;t=realloc(t,(size_t)cap);if(!t)return NULL;}
   n+=snprintf(t+n,cap-(size_t)n,"lab_booking_request_latency_ms_bucket{le=\"%s\"} %lld\n",le,(long long)cum);
  }
- if((size_t)n+320>=(int)cap){cap*=2;t=realloc(t,(size_t)cap);if(!t)return NULL;}
+ if((size_t)n+320>=(size_t)cap){cap*=2;t=realloc(t,(size_t)cap);if(!t)return NULL;}
  n+=snprintf(t+n,cap-(size_t)n,
   "lab_booking_request_latency_ms_sum %lld\n"
   "lab_booking_request_latency_ms_count %lld\n",
